\name{rsnot-package}
\alias{rsnot-package}
\alias{rsnot}
\docType{package}
\title{
\packageTitle{rsnot}
}
\description{
\packageDescription{rsnot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rsnot}
\packageIndices{rsnot}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rsnot}

Maintainer: \packageMaintainer{rsnot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
