% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnot.R
\name{snot_dump}
\alias{snot_dump}
\title{Dump a SNOT object to a string.}
\usage{
snot_dump(snot_object)
}
\arguments{
\item{snot_object}{A hierarchical R object built of basic R datatypes to be dumped to SNOT/JSON.}
}
\value{
A string representation of the data provided.
}
\description{
Dump a SNOT object to a string.
}
\examples{
snot_dump(c(1,2,3,4))
}
