% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnot.R
\name{snot_load}
\alias{snot_load}
\title{Load a SNOT object from a string.}
\usage{
snot_load(snot_string)
}
\arguments{
\item{snot_string}{A string containing a complete SNOT/JSON object.}
}
\value{
An object built of basic R types reflecting the structure of the SNOT data in the string.
}
\description{
Load a SNOT object from a string.
}
\examples{
snot_load('{"a":1.0, "b":[2.0,3.0,{"c":3.14159}]}')
}
