% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnot.R
\name{snot_populate}
\alias{snot_populate}
\title{Assign values to a SNOT object from values in a vector.}
\usage{
snot_populate(the_structure, a_vector)
}
\arguments{
\item{the_structure}{A hierarchical R object built of basic R datatypes (A SNOT object).}

\item{a_vector}{An ordered vector with all of the values to assign to leaves in the structure.}
}
\value{
A SNOT object matching the structure of the_structure, but with leaves having values taken from a_vector.
}
\description{
Assign values to a SNOT object from values in a vector.
}
