% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsnot.R
\name{snot_traverse}
\alias{snot_traverse}
\title{Get all the leaf values into a vector.}
\usage{
snot_traverse(the_structure)
}
\arguments{
\item{the_structure}{A hierarchical R object built of basic R datatypes (A SNOT object).}
}
\value{
A vector with all of the ordered values of the leaves of the_structure.
}
\description{
Get all the leaf values into a vector.
}
